+++ tests/integration/api/v2010/account/sip/test_domain.py
@@ -54,7 +54,7 @@ class DomainTestCase(IntegrationTestCase):
                         "voice_status_callback_method": "POST",
                         "voice_status_callback_url": null,
                         "voice_url": "https://dundermifflin.example.com/twilio/app.php",
-                        "trunk_sid": "TRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                        "emergency_calling_enabled": true
                     }
                 ],
                 "start": 0,
@@ -140,7 +140,7 @@ class DomainTestCase(IntegrationTestCase):
                 "voice_status_callback_method": "POST",
                 "voice_status_callback_url": null,
                 "voice_url": "https://dundermifflin.example.com/twilio/app.php",
-                "trunk_sid": null
+                "emergency_calling_enabled": true
             }
             '''
         ))
@@ -189,7 +189,7 @@ class DomainTestCase(IntegrationTestCase):
                 "voice_status_callback_method": "POST",
                 "voice_status_callback_url": null,
                 "voice_url": "https://dundermifflin.example.com/twilio/app.php",
-                "trunk_sid": "TRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                "emergency_calling_enabled": true
             }
             '''
         ))
@@ -238,7 +238,7 @@ class DomainTestCase(IntegrationTestCase):
                 "voice_status_callback_method": "POST",
                 "voice_status_callback_url": null,
                 "voice_url": "https://dundermifflin.example.com/twilio/app.php",
-                "trunk_sid": "TRaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                "emergency_calling_enabled": true
             }
             '''
         ))