+++ twilio/rest/api/v2010/account/call/__init__.py
@@ -60,7 +60,8 @@ class CallList(ListResource):
                machine_detection_silence_timeout=values.unset,
                async_amd=values.unset, async_amd_status_callback=values.unset,
                async_amd_status_callback_method=values.unset, byoc=values.unset,
-               url=values.unset, twiml=values.unset, application_sid=values.unset):
+               call_reason=values.unset, url=values.unset, twiml=values.unset,
+               application_sid=values.unset):
         """
         Create the CallInstance
 
@@ -92,6 +93,7 @@ class CallList(ListResource):
         :param unicode async_amd_status_callback: The URL we should call to send amd status information to your application
         :param unicode async_amd_status_callback_method: HTTP Method to use with async_amd_status_callback
         :param unicode byoc: BYOC trunk SID (Beta)
+        :param unicode call_reason: Reason for the call (Beta)
         :param unicode url: The absolute URL that returns TwiML for this call
         :param unicode twiml: TwiML instructions for the call
         :param unicode application_sid: The SID of the Application resource that will handle the call
@@ -131,6 +133,7 @@ class CallList(ListResource):
             'AsyncAmdStatusCallback': async_amd_status_callback,
             'AsyncAmdStatusCallbackMethod': async_amd_status_callback_method,
             'Byoc': byoc,
+            'CallReason': call_reason,
         })
 
         payload = self._version.create(method='POST', uri=self._uri, data=data, )